requirements.txt
pandas>=1.3
numpy>=1.21
pyyaml>=6.0
scipy>=1.7
scikit-learn>=1.0
python-dateutil
tqdm
